[TWINFACES-324] refactor: change folder name
