+++ src/app/workspace/pipeline-steps/page.tsx
@@ -1,4 +1,4 @@
-import { PipelineStepsScreen } from "@/screens/pipelineSteps";
+import { PipelineStepsScreen } from "@/screens/pipeline-steps";
 
 export default function PipelineStepsPage() {
   return <PipelineStepsScreen />;
